# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My Product API Project",
        "description": "My Product API Project",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:2000/",
    "basePath": "/",
    "tags": [
        {
            "name": "Products",
            "description": "API for products in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/allproducts": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products in system",                
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/allproducts?page={page}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated all products in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/allproducts?page={page}&limit={limit}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated all products by limit in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "description": "Limit that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/bestSellers": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Best Seller Products in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/bestSellers?page={page}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Paginated Best Seller Products in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/bestSellers?page={page}&limit={limit}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Paginated Best Seller Products by limit in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "description": "Limit that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/electronic": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Electronic Products in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/electronic?page={page}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Paginated Electronic Products in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/electronic?page={page}&limit={limit}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get Paginated Electronic Products by limit in system",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "description": "Page that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "description": "Limit that we want to get",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products?q={q}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search all products in system",
                "parameters": [
                    {
                        "name": "q",
                        "in": "path",
                        "required": true,
                        "description": "Search Query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by id in system",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Product Id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Save Product",
                "parameters": [
                    {
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "description": "Product Name",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/update/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "parameters": [
                    {
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "description": "Product Name",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Product Id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        },
        "/products/delete/{id}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete Product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Product Id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Products"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Product": {
            "required": [
                "name",
                "price",
                "productType"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "productType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "Products": {
            "type": "array",
            "$ref": "#/definitions/Product"
        }
    }
}